ZQS-671 move protocols into interfaces subpackage
